requirements.txt
Flask
flask
openai
python-dotenv
pymongo
google-api-python-client
google-auth
google-auth-oauthlib
apscheduler
twilio
deep-translator
langdetect
pandas
requests